Function test_generation::transitions::stack_local_polymorphic_eq

pub fn stack_local_polymorphic_eq(
    state: &AbstractState, 
    index1: usize, 
    index2: usize
) -> bool

Determine whether an abstract value on the stack and a abstract value in the locals have the same type